# docs/swagger.yaml
basePath: /
definitions:
  handler.ErrorResponse:
    properties:
      details: {}
      error:
        example: bad_request
        type: string
      message:
        example: invalid request body
        type: string
      request_id:
        example: 3f2b1c9d8e7a6b5c4d3e2f1a0b9c8d7e
        type: string
    type: object
  handler.HealthCheckResponse:
    properties:
      checked:
        type: string
      duration:
        example: 3
        type: integer
      status:
        $ref: '#/definitions/handler.HealthStatus'
    type: object
  handler.HealthStatus:
    properties:
      database:
        example: up
        type: string
      redis_cache:
        example: up
        type: string
    type: object
  handler.SuccessResponse:
    properties:
      data: {}
      message:
        example: user created
        type: string
      success:
        example: true
        type: boolean
    type: object
  handler.UserResponse:
    properties:
      clerk_id:
        example: user_2abcXYZ
        type: string
      created_at:
        example: "2026-10-05T12:00:00Z"
        type: string
      email:
        example: jane@example.com
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        example: Jane Doe
        type: string
      updated_at:
        example: "2026-10-05T12:00:00Z"
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Dockerized Go API starter using Echo, PostgreSQL, sqlc and Redis.
  title: Boiler-Go API
  version: "1.0"
paths:
  /health:
    get:
      description: Returns database and Redis cache status. Returns 503 when the database
        is down.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.HealthCheckResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/handler.HealthCheckResponse'
      summary: Health check
      tags:
      - health
  /me:
    get:
      description: Returns the local user linked to the Clerk session. Requires a
        Clerk session token.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Current user
      tags:
      - users
  /repo-test/users:
    get:
      description: List users newest-first. Requires a Clerk session token.
      parameters:
      - default: 10
        description: Max rows
        in: query
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.UserResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - users
  /repo-test/users/get:
    get:
      description: Get a single user by id query param. Requires a Clerk session token.
      parameters:
      - description: User UUID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user
      tags:
      - users
  /webhooks/clerk:
    post:
      consumes:
      - application/json
      description: Receives Clerk user lifecycle events (Svix-signed) and syncs local
        users.
      parameters:
      - description: Svix message ID
        in: header
        name: svix-id
        required: true
        type: string
      - description: Svix timestamp
        in: header
        name: svix-timestamp
        required: true
        type: string
      - description: Svix signature
        in: header
        name: svix-signature
        required: true
        type: string
      - description: Clerk event payload
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Clerk webhook
      tags:
      - webhooks
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and your Clerk session token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"